Renamed
"Timestamp tree link" didn't sound right.
"Time Machine" here should remind of macOS' Time Machine or just "going back in time"
Name suggestions welcome
